requirements.txt
pip-tools
justai==3.2.10
python-dotenv
tiktoken
lxml
justdays